+++ packages/types/src/types/seasonal/2021.ts
@@ -3,6 +3,20 @@ import { TypeHidden, TypeState, TypeTags } from "../../munzee";
 
 const s2021: SeasonalCategory[] = [
   {
+    name: "Zee Folk",
+    id: "zeefolk2020",
+    starts: "2021-03-17T00:00:00-05:00",
+    ends: "2021-03-31T23:59:00-05:00",
+    pobs: [
+      {
+        name: "Zee Folk",
+        icon: "zeefolk",
+        id: "null_zeefolk",
+        lands_on: [],
+      },
+    ],
+  },
+  {
     name: "Women's History Month",
     id: "womenshistory2020",
     starts: "2021-02-25T12:00:00-05:00",